Split file parsing into own file
